Ensure correct phase ID from ANG reader, fix pg aliases [skip ci]
